backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic[email]==2.7.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
pgvector==0.2.5

langdetect==1.0.9
openai>=1.0.0

numpy==1.26.4
argparse==1.4.0
httpx==0.27.0
redis==5.0.3
structlog==24.1.0
python-multipart==0.0.9
tenacity==8.2.3
boto3==1.34.84
stripe==9.2.0
sentry-sdk==1.45.0
uvloop==0.19.0

# Google Cloud APIs for OCR/Document processing
google-cloud-vision==3.7.2
google-cloud-documentai==2.26.0
google-api-python-client==2.118.0
Pillow==10.2.0

transformers==4.40.1
torch>=2.5,<2.6
sentencepiece==0.2.0
sentence-transformers==2.7.0
epitran==1.24
simalign

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
httpx==0.27.0
pytest-mock==3.12.0

# Observability
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
prometheus-client==0.20.0
